requirements.txt
numpy==1.25.2
Python == 3.10.9
pandas>=2.0.3
python-dotenv
Flask
APScheduler
plotly
yfinance